# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "API Documentation for User Management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/certificates": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Get all certificates",
                "operationId": "542620d99fb7a93d6889bcd434d9afe0",
                "responses": {
                    "200": {
                        "description": "List of certificates"
                    },
                    "404": {
                        "description": "No certificates found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Create a new certificate",
                "operationId": "4e782d44d2eb7763463a7a56e77545a0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "course_id",
                                    "certificate_path"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "issued_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-23"
                                    },
                                    "certificate_path": {
                                        "type": "string",
                                        "example": "/certificates/cert_123.pdf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Certificate created successfully"
                    }
                }
            }
        },
        "/api/certificates/{id}": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Get certificate by ID",
                "operationId": "191e2a30f6ace09ac91b4b07289cfa86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certificate details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Update a certificate",
                "operationId": "491a76b1ff829a5822f5829f94d3c35f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "issued_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-23"
                                    },
                                    "certificate_path": {
                                        "type": "string",
                                        "example": "/certificates/cert_123_updated.pdf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Certificate updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Delete a certificate",
                "operationId": "733beb8b840fc22a4506f6a8d59784c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certificate deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}